// web/src/api/generated/apis/EntitlementsApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime';
import {
    type ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';
import {
    type SpaceEntitlementView,
    SpaceEntitlementViewFromJSON,
    SpaceEntitlementViewToJSON,
} from '../models/SpaceEntitlementView';

export interface GetSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRequest {
    spaceId: string;
}

/**
 * 
 */
export class EntitlementsApi extends runtime.BaseAPI {

    /**
     * Creates request options for getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGet without sending the request
     */
    async getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRequestOpts(requestParameters: GetSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['spaceId'] == null) {
            throw new runtime.RequiredError(
                'spaceId',
                'Required parameter "spaceId" was null or undefined when calling getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGet().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};


        let urlPath = `/api/v1/spaces/{spaceId}/entitlements`;
        urlPath = urlPath.replace('{spaceId}', encodeURIComponent(String(requestParameters['spaceId'])));

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     * Return the effective commercial capability entitlement state for the Space.
     * Get Space Entitlements
     */
    async getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRaw(requestParameters: GetSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<SpaceEntitlementView>> {
        const requestOptions = await this.getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => SpaceEntitlementViewFromJSON(jsonValue));
    }

    /**
     * Return the effective commercial capability entitlement state for the Space.
     * Get Space Entitlements
     */
    async getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGet(requestParameters: GetSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<SpaceEntitlementView> {
        const response = await this.getSpaceEntitlementsApiV1SpacesSpaceIdEntitlementsGetRaw(requestParameters, initOverrides);
        return await response.value();
    }

}