// web/src/api/generated/models/DailyQuoteCatalogView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { QuoteSourceView } from './QuoteSourceView';
import {
    QuoteSourceViewFromJSON,
    QuoteSourceViewFromJSONTyped,
    QuoteSourceViewToJSON,
    QuoteSourceViewToJSONTyped,
} from './QuoteSourceView';
import type { QuoteCategoryView } from './QuoteCategoryView';
import {
    QuoteCategoryViewFromJSON,
    QuoteCategoryViewFromJSONTyped,
    QuoteCategoryViewToJSON,
    QuoteCategoryViewToJSONTyped,
} from './QuoteCategoryView';

/**
 * 
 * @export
 * @interface DailyQuoteCatalogView
 */
export interface DailyQuoteCatalogView {
    /**
     * 
     * @type {Array<QuoteCategoryView>}
     * @memberof DailyQuoteCatalogView
     */
    categories: Array<QuoteCategoryView>;
    /**
     * 
     * @type {Array<QuoteSourceView>}
     * @memberof DailyQuoteCatalogView
     */
    sources: Array<QuoteSourceView>;
}

/**
 * Check if a given object implements the DailyQuoteCatalogView interface.
 */
export function instanceOfDailyQuoteCatalogView(value: object): value is DailyQuoteCatalogView {
    if (!('categories' in value) || value['categories'] === undefined) return false;
    if (!('sources' in value) || value['sources'] === undefined) return false;
    return true;
}

export function DailyQuoteCatalogViewFromJSON(json: any): DailyQuoteCatalogView {
    return DailyQuoteCatalogViewFromJSONTyped(json, false);
}

export function DailyQuoteCatalogViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyQuoteCatalogView {
    if (json == null) {
        return json;
    }
    return {
        
        'categories': ((json['categories'] as Array<any>).map(QuoteCategoryViewFromJSON)),
        'sources': ((json['sources'] as Array<any>).map(QuoteSourceViewFromJSON)),
    };
}

export function DailyQuoteCatalogViewToJSON(json: any): DailyQuoteCatalogView {
    return DailyQuoteCatalogViewToJSONTyped(json, false);
}

export function DailyQuoteCatalogViewToJSONTyped(value?: DailyQuoteCatalogView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'categories': ((value['categories'] as Array<any>).map(QuoteCategoryViewToJSON)),
        'sources': ((value['sources'] as Array<any>).map(QuoteSourceViewToJSON)),
    };
}