// web/src/api/generated/models/DailyQuotePreferenceView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DailyQuotePreferenceView
 */
export interface DailyQuotePreferenceView {
    /**
     * 
     * @type {string}
     * @memberof DailyQuotePreferenceView
     */
    accountId: string;
    /**
     * 
     * @type {boolean}
     * @memberof DailyQuotePreferenceView
     */
    enabled: boolean;
    /**
     * 
     * @type {string}
     * @memberof DailyQuotePreferenceView
     */
    locale?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof DailyQuotePreferenceView
     */
    selectedCategoryIds: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DailyQuotePreferenceView
     */
    selectedSourceIds: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof DailyQuotePreferenceView
     */
    version: number;
}

/**
 * Check if a given object implements the DailyQuotePreferenceView interface.
 */
export function instanceOfDailyQuotePreferenceView(value: object): value is DailyQuotePreferenceView {
    if (!('accountId' in value) || value['accountId'] === undefined) return false;
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    if (!('selectedCategoryIds' in value) || value['selectedCategoryIds'] === undefined) return false;
    if (!('selectedSourceIds' in value) || value['selectedSourceIds'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function DailyQuotePreferenceViewFromJSON(json: any): DailyQuotePreferenceView {
    return DailyQuotePreferenceViewFromJSONTyped(json, false);
}

export function DailyQuotePreferenceViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyQuotePreferenceView {
    if (json == null) {
        return json;
    }
    return {
        
        'accountId': json['accountId'],
        'enabled': json['enabled'],
        'locale': json['locale'] === undefined ? undefined : json['locale'] === null ? null : json['locale'],
        'selectedCategoryIds': json['selectedCategoryIds'],
        'selectedSourceIds': json['selectedSourceIds'],
        'version': json['version'],
    };
}

export function DailyQuotePreferenceViewToJSON(json: any): DailyQuotePreferenceView {
    return DailyQuotePreferenceViewToJSONTyped(json, false);
}

export function DailyQuotePreferenceViewToJSONTyped(value?: DailyQuotePreferenceView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'accountId': value['accountId'],
        'enabled': value['enabled'],
        'locale': value['locale'],
        'selectedCategoryIds': value['selectedCategoryIds'],
        'selectedSourceIds': value['selectedSourceIds'],
        'version': value['version'],
    };
}