// web/src/api/generated/models/DailyQuoteResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DailyQuoteView } from './DailyQuoteView';
import {
    DailyQuoteViewFromJSON,
    DailyQuoteViewFromJSONTyped,
    DailyQuoteViewToJSON,
    DailyQuoteViewToJSONTyped,
} from './DailyQuoteView';

/**
 * 
 * @export
 * @interface DailyQuoteResponse
 */
export interface DailyQuoteResponse {
    /**
     * 
     * @type {Date}
     * @memberof DailyQuoteResponse
     */
    checkedOn: Date;
    /**
     * 
     * @type {boolean}
     * @memberof DailyQuoteResponse
     */
    enabled: boolean;
    /**
     * 
     * @type {DailyQuoteView}
     * @memberof DailyQuoteResponse
     */
    quote?: DailyQuoteView | null;
}

/**
 * Check if a given object implements the DailyQuoteResponse interface.
 */
export function instanceOfDailyQuoteResponse(value: object): value is DailyQuoteResponse {
    if (!('checkedOn' in value) || value['checkedOn'] === undefined) return false;
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    return true;
}

export function DailyQuoteResponseFromJSON(json: any): DailyQuoteResponse {
    return DailyQuoteResponseFromJSONTyped(json, false);
}

export function DailyQuoteResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyQuoteResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'checkedOn': (new Date(json['checkedOn'])),
        'enabled': json['enabled'],
        'quote': json['quote'] === undefined ? undefined : json['quote'] === null ? null : DailyQuoteViewFromJSON(json['quote']),
    };
}

export function DailyQuoteResponseToJSON(json: any): DailyQuoteResponse {
    return DailyQuoteResponseToJSONTyped(json, false);
}

export function DailyQuoteResponseToJSONTyped(value?: DailyQuoteResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'checkedOn': value['checkedOn'].toISOString().substring(0,10),
        'enabled': value['enabled'],
        'quote': DailyQuoteViewToJSON(value['quote']),
    };
}